Added excel import and export, currently it only populates 1 cell. Noted issue with adding new purchase -> products can only be added during modify.

+++ src/main/java/hu/torma/deliveryapplication/entity/Product.java
@@ -22,10 +22,10 @@ public class Product {
     private Unit secondUnit;
 
     @Column(name = "price")
-    private Long price;
+    private Double price;
 
     @Column(name = "compensational")
-    private long compPercent;
+    private Double compPercent;
 
     @Column(name = "tariffnumber")
     private String tariffnum;
